+++ CharlieBackend.Api/Controllers/StudentsController.cs
@@ -101,6 +101,22 @@ namespace CharlieBackend.Api.Controllers
         }
 
         /// <summary>
+        /// Gets all of the student's study group
+        /// </summary>
+        /// <response code="200">Successful return of student's study groups</response>
+        /// <response code="HTTP: 404, API: 3">Error, can not find student or student's study groups</response>
+        [SwaggerResponse(200, type: typeof(IList<StudentStudyGroupsDto>))]
+        [Authorize(Roles = "Secretary, Mentor, Admin")]
+        [HttpGet("{id}/groups")]
+        public async Task<ActionResult<IList<StudentStudyGroupsDto>>> GetStudentStudyGroupsByStudentId(long id)
+        {
+            var foundGroups = await _studentService
+                    .GetStudentStudyGroupsByStudentIdAsync(id);
+
+            return foundGroups.ToActionResult();
+        }
+
+        /// <summary>
         /// Updates student
         /// </summary>
         /// <response code="200">Successful update of student</response>
