+++ CharlieBackend.Api/SwaggerExamples/StudentsController/StudentStudyGroupsResponse.cs
@@ -1,11 +1,12 @@
-﻿using CharlieBackend.Core.DTO.StudentGroups;
+﻿using CharlieBackend.Core.DTO.Student;
+using CharlieBackend.Core.DTO.StudentGroups;
 using Swashbuckle.AspNetCore.Filters;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace CharlieBackend.Api.SwaggerExamples.StudentGroupsController
+namespace CharlieBackend.Api.SwaggerExamples.StudentsController
 {
     class StudentStudyGroupsResponse : IExamplesProvider<IList<StudentStudyGroupsDto>>
     {
