+++ CharlieBackend.Business/Services/StudentGroupService.cs
@@ -193,22 +193,5 @@ namespace CharlieBackend.Business.Services
 
             return Result<StudentGroupDto>.GetSuccess(_mapper.Map<StudentGroupDto>(foundStudentGroup));
         }
-
-        public async Task<Result<IList<StudentStudyGroupsDto>>> GetStudentStudyGroupsByStudentIdAsync(long id)
-        {
-            var foundGroups = await _unitOfWork.StudentGroupRepository.GetStudentStudyGroups(id);
-
-            if (!_unitOfWork.StudentRepository.IsStudentExist(id))
-            {
-                return Result<IList<StudentStudyGroupsDto>>.GetError(ErrorCode.NotFound, "Student doesn`t exist");
-            }
-
-            if (!foundGroups.Any())
-            {
-                return Result<IList<StudentStudyGroupsDto>>.GetError(ErrorCode.NotFound, $"Study groups for student with id {id} not found");
-            }
-
-            return Result<IList<StudentStudyGroupsDto>>.GetSuccess(foundGroups);
-        }
     }
 }
